runpod-worker/requirements.txt
torch>=2.1.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors
sentencepiece
Pillow
redis
boto3
